@@ -1,9 +1,16 @@
 {
   "name": "@warerebel/azureblobstoragehelper",
-  "version": "0.7.0",
+  "version": "0.8.0",
   "description": "Module to ease the storing of data in azure gen2 blob storage",
-  "main": "lib/azureBlobStorage.js",
+  "main": "dist/azureBlobStorage.js",
+  "files": [
+    "dist/*.js",
+    "src/*.ts",
+    "test/*.spec.js",
+    "index.d.ts"
+  ],
   "scripts": {
+    "build": "node node_modules/typescript/bin/tsc",
     "test": "node node_modules/nyc/bin/nyc node_modules/mocha/bin/mocha test/*.spec.js",
     "lint": "node node_modules/eslint/bin/eslint **/*.js",
     "coverage": "node node_modules/nyc/bin/nyc report --reporter=text-lcov | node_modules/coveralls/bin/coveralls.js"
@@ -25,13 +32,17 @@
   },
   "homepage": "https://github.com/warerebel/azureblobstoragehelper#readme",
   "dependencies": {
-    "@warerebel/azurerestauth": "^0.2.1"
+    "@warerebel/azurerestauth": "^0.4.9"
   },
   "devDependencies": {
+    "@types/node": "^12.7.8",
+    "@typescript-eslint/eslint-plugin": "^2.3.1",
+    "@typescript-eslint/parser": "^2.3.1",
     "coveralls": "^3.0.6",
-    "eslint": "^6.3.0",
+    "eslint": "^6.4.0",
     "mocha": "^6.2.0",
     "nyc": "^14.1.1",
-    "sinon": "^7.4.2"
+    "sinon": "^7.5.0",
+    "typescript": "^3.6.3"
   }
 }
